requirements.txt
fastapi
uvicorn[standard]
azure-identity
openai
requests
python-dotenv
-e ./llm-reasoning-api/api/algorithms/llm-reasoners